+++ ImageFeed/Auth/WebViewViewController.swift
@@ -8,6 +8,11 @@ public protocol WebViewViewControllerProtocol: AnyObject {
     func setProgressHidden(_ isHidden: Bool)
 }
 
+protocol WebViewViewControllerDelegateProtocol: AnyObject {
+    func webViewViewController(_ vc: WebViewViewController, didAuthenticateWithCode code: String)
+    func webViewViewControllerDidCancel(_ vc: WebViewViewController)
+}
+
 final class WebViewViewController: UIViewController & WebViewViewControllerProtocol {
 
     //MARK: - Public Properties
