--- ImageFeed/Auth/WebViewViewControllerDelegateProtocol.swift
@@ -1,4 +1,0 @@
-protocol WebViewViewControllerDelegateProtocol: AnyObject {
-    func webViewViewController(_ vc: WebViewViewController, didAuthenticateWithCode code: String)
-    func webViewViewControllerDidCancel(_ vc: WebViewViewController)
-}
